More renaming
Still thinking about what to do with fit_multiple, but this should be most of the file renaming necessary for fit_resonator module
