+++ fit_resonator/test/fit_S_data_test.py
@@ -1,4 +1,4 @@
-import fit_resonator.fit_S_data as fsd
+import fit_resonator.Sdata as fsd
 import numpy as np
 import pytest
