+++ fit_resonator/test/user_testing.py
@@ -9,7 +9,7 @@ from scipy.interpolate import interp1d
 pathToParent = os.path.dirname(os.path.dirname(os.path.abspath(__file__))) #set a variable that equals the relative path of parent directory
 sys.path.append(pathToParent)#path to Fit_Cavity
 import fit_resonator.resonator as res
-import fit_resonator.fit_S_data as fsd
+import fit_resonator.Sdata as fsd
 np.set_printoptions(precision=4,suppress=True)# display numbers with 4 sig. figures (digits)
 
                          ## Code Starts Here ##
@@ -23,8 +23,8 @@ filename = 'R~10^6.txt'
 fit_type = 'DCM'
 MC_iteration = 10
 MC_rounds = 1e3
-MC_fix = ['Q','Qi','Qc','Qa','phi','w1']
-#MC_fix = []
+#MC_fix = ['Q','Qi','Qc','Qa','phi','w1']
+MC_fix = []
 manual_init = [10000,100000,5.797,-0.558561]        #make your own initial guess: [Qi, Qc, freq, phi] (instead of phi used Qa for CPZM)
 #manual_init = None # find initial guess by itself
 
@@ -40,7 +40,7 @@ Method = res.FitMethod(fit_type, MC_iteration, MC_rounds=MC_rounds,\
 normalize = 10
 
 ### Fit Resonator function without background removal ###
-params,conf_array,fig1,chi1,init1 = fsd.fit_resonator(filename = filename,Method = Method,normalize = normalize,dir = dir,preprocess_method = "circle")
+params,conf_array,fig1,chi1,init1 = fsd.fit(filename = filename, Method = Method, normalize = normalize, dir = dir, preprocess_method ="circle")
 
 ### Fit Resonator function with background removal ###
 #background_file = 'example_background.csv'
